refactor: Move configuration classes into Configuration subdirectory
Moving classes in the BinaryInjection module to folders to make reading the code easier.
