+++ .gitignore
@@ -138,6 +138,9 @@ DocProject/Help/html
 # Click-Once directory
 publish/
 
+# Binary Dependencies
+deps/
+
 # Publish Web Output
 *.[Pp]ublish.xml
 *.azurePubxml
