--- src/CoreHook.BinaryInjection/Loader/ClrHostConfiguration.Windows.cs
@@ -1,9 +1,0 @@
-﻿
-namespace CoreHook.BinaryInjection.Loader
-{
-    public sealed partial class ClrHostConfiguration
-    {
-        public static string ClrStartFunction => "StartCoreCLR";
-        public static string ClrExecuteManagedFunction => "ExecuteAssemblyFunction";
-    }
-}
