+++ src/CoreHook.BinaryInjection/Loader/Configuration/ClrHostConfiguration.cs
@@ -1,0 +1,9 @@
+﻿
+namespace CoreHook.BinaryInjection.Loader.Configuration
+{
+    internal sealed class ClrHostConfiguration
+    {
+        internal static string ClrStartFunction => "StartCoreCLR";
+        internal static string ClrExecuteManagedFunction => "ExecuteAssemblyFunction";
+    }
+}
