+++ src/CoreHook.BinaryInjection/Loader/Configuration/IStringEncodingConfiguration.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text;
 
-namespace CoreHook.BinaryInjection.Loader
+namespace CoreHook.BinaryInjection.Loader.Configuration
 {
     public interface IStringEncodingConfiguration
     {
