+++ src/CoreHook.BinaryInjection/Loader/HostFunctionArguments.cs
@@ -1,4 +1,5 @@
 ﻿using System.IO;
+using CoreHook.BinaryInjection.Loader.Configuration;
 using CoreHook.BinaryInjection.Loader.Serialization;
 
 namespace CoreHook.BinaryInjection.Loader
