+++ src/CoreHook.BinaryInjection/RemoteInjection/RemoteInjector.cs
@@ -5,7 +5,9 @@ using System.IO;
 using System.Reflection;
 using System.Runtime.InteropServices;
 using CoreHook.BinaryInjection.Loader;
-using CoreHook.BinaryInjection.ProcessUtils;
+using CoreHook.BinaryInjection.Loader.Configuration;
+using CoreHook.BinaryInjection.ProcessUtils.Configuration;
+using CoreHook.BinaryInjection.RemoteInjection.Configuration;
 using CoreHook.CoreLoad.Data;
 using CoreHook.IPC.Platform;
 using CoreHook.Memory.Processes;
